test: test default-layout redirection on missing session or user
